+++ tests/ops/integration_tests/saas/test_datadog_task.py
@@ -21,6 +21,7 @@ def test_datadog_connection_test(datadog_connection_config) -> None:
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_datadog
+@pytest.mark.asyncio
 async def test_saas_access_request_task(
     db,
     policy,