+++ tests/ops/integration_tests/saas/test_sendgrid_task.py
@@ -15,6 +15,7 @@ from tests.ops.test_helpers.saas_test_utils import poll_for_existence
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_sendgrid
+@pytest.mark.asyncio
 async def test_sendgrid_access_request_task(
     db,
     policy,
@@ -69,6 +70,7 @@ async def test_sendgrid_access_request_task(
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_sendgrid
+@pytest.mark.asyncio
 async def test_sendgrid_erasure_request_task(
     db,
     policy,