+++ tests/ops/integration_tests/saas/test_sentry_task.py
@@ -18,6 +18,7 @@ from tests.ops.test_helpers.saas_test_utils import poll_for_existence
 @pytest.mark.skip(reason="Pending account resolution")
 @pytest.mark.integration_saas
 @pytest.mark.integration_sentry
+@pytest.mark.asyncio
 async def test_sentry_access_request_task(
     db,
     policy,
@@ -265,6 +266,7 @@ def sentry_erasure_test_prep(sentry_connection_config, db):
 @pytest.mark.skip(reason="Pending account resolution")
 @pytest.mark.integration_saas
 @pytest.mark.integration_sentry
+@pytest.mark.asyncio
 async def test_sentry_erasure_request_task(
     db, policy, sentry_connection_config, sentry_dataset_config
 ) -> None: