+++ tests/ops/integration_tests/saas/test_stripe_task.py
@@ -16,6 +16,7 @@ from tests.ops.graph.graph_test_util import assert_rows_match
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_stripe
+@pytest.mark.asyncio
 async def test_stripe_access_request_task(
     db,
     policy,
@@ -634,6 +635,7 @@ async def test_stripe_access_request_task(
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_stripe
+@pytest.mark.asyncio
 async def test_stripe_erasure_request_task(
     db,
     policy,